requirements.txt
tkinter
Pillow
pyperclip
groq
spotipy
SpeechRecognition
pyaudio
requests
pywin32
